Refactor the nav button
