--- src/components/Button/ButtonFunction/styled.ts
@@ -1,3 +1,0 @@
-import styled from 'styled-components'
-
-export const ButtonCover = styled.div``
