--- src/components/Button/ButtonLink/styled.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-import { Link } from 'react-router-dom'
-
-export const StyledLink = styled(Link)`
-	text-decoration: none;
-`
